Handouts-Textbook-The Hardware of a Quantum Computer_Part_2
This course aims to give you an understanding of the scientific basis behind the quantum computer and a quantum internet. Our challenge to you is to really understand the working principles of the different layers of a quantum computer and a quantum internet. This course is about getting an insight into the electronics, hardware and software that are essential for a practical quantum computer and a quantum internet.